[Runtime] Fix CRaC failed jtreg testcases and support load CRaC classses lazily.
Summary:  Most of the CRaC testcases run failed when run on the lastest criu and linux kernel 5.10. The first is write to file instead of stdout and stderr, because stdout and stderr depend on pipe or tty, but pipe and tty cannot be checkpoint and restore correctly with criu. The second is use 'docker run' instead of 'docker exec' to run java application in docker.If checkpoint a application that lauched by 'docker exec', there is an error "Can't lookup mount=24 for fd=0 path=/dev/null".

Testing: All testcases in jdk/jdk/crac.

Reviewers: lei.yul, denghui.ddh

Issue: https://github.com/dragonwell-project/dragonwell11/issues/867