[Runtime] CRaC: Restore stdout and stderr when run CRaC in unprivileged docker.
Summary: The stdout and stderr are pipe files when run in docker, restore these pipe files is rather tricky. Write the fds to a file named pipefds, then after criu checkpoint successfully, the criu execute the criuengine as a postdump callback. In the callback, append the pipe info of java process to the file pipefds. It read the pipefds when restore, than pass the pipe info as --inherit-fd to criu. The problem is criuengine cannot get the pipefds file path if run with nonprivilged. To fix this, set the environment CRAC_IMAGE_DIR explictly when do checkpointing.

Testing: jdk/jdk/crac/stdoutInDocker/TestStdoutInDocker.sh

Reviewers: lei.yul,denghui.ddh

Issue: https://github.com/dragonwell-project/dragonwell11/issues/867